CashFlow--Business-decision-support-system-main/backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
motor==3.6.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pymongo==4.9.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
pydantic-settings==2.1.0
pyzbar>=0.1.9
opencv-python>=4.5.0
numpy>=1.20.0
# Sales prediction dependencies
pandas==2.0.0
numpy==1.24.3
scikit-learn==1.2.2
joblib==1.2.0
matplotlib>=3.4.0
seaborn>=0.11.0
